+++ dd-java-agent/instrumentation-testing/src/main/groovy/datadog/trace/agent/test/SpockExtension.java
@@ -1,6 +1,8 @@
 package datadog.trace.agent.test;
 
 import java.io.IOException;
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 import java.util.List;
 import net.bytebuddy.dynamic.ClassFileLocator;
 import org.junit.jupiter.api.extension.AfterAllCallback;
@@ -20,16 +22,40 @@ public final class SpockExtension
         TestExecutionExceptionHandler {
 
   private static final ExtensionContext.Namespace NAMESPACE =
-      ExtensionContext.Namespace.create("dd", "spock", "loader");
+      ExtensionContext.Namespace.create("dd", "spock");
+
+  private static final String INSTRUMENTATION_CLASSLOADER = "instrumentation-class-loader";
+  private static final String ORIGINAL_CLASSLOADER = "original-thread-context-class-loader";
 
   @Override
   public void beforeAll(ExtensionContext ctx) {
     Class<?> testClass = ctx.getRequiredTestClass();
-    BootstrapClasspathSetup.assertNoBootstrapClassesInTestClass(testClass);
+    assertNoBootstrapClassesInTestClass(testClass);
 
     InstrumentationClassLoader custom =
         new InstrumentationClassLoader(testClass.getClassLoader(), testClass.getName());
-    ctx.getStore(NAMESPACE).put("loader", custom);
+    ctx.getStore(NAMESPACE).put(INSTRUMENTATION_CLASSLOADER, custom);
+  }
+
+  public static void assertNoBootstrapClassesInTestClass(final Class<?> testClass) {
+    for (final Field field : testClass.getDeclaredFields()) {
+      assertNotBootstrapClass(testClass, field.getType());
+    }
+    for (final Method method : testClass.getDeclaredMethods()) {
+      assertNotBootstrapClass(testClass, method.getReturnType());
+      for (final Class<?> paramType : method.getParameterTypes()) {
+        assertNotBootstrapClass(testClass, paramType);
+      }
+    }
+  }
+
+  private static void assertNotBootstrapClass(final Class<?> testClass, final Class<?> clazz) {
+    if (!clazz.isPrimitive() && BootstrapClasspathSetup.isBootstrapClass(clazz.getName())) {
+      throw new IllegalStateException(
+          testClass.getName()
+              + ": Bootstrap classes are not allowed in test class field or method signatures. Offending class: "
+              + clazz.getName());
+    }
   }
 
   @Override
@@ -39,14 +65,16 @@ public final class SpockExtension
 
   @Override
   public void beforeEach(ExtensionContext ctx) {
-    ClassLoader custom = ctx.getStore(NAMESPACE).get("loader", ClassLoader.class);
-    ctx.getStore(NAMESPACE).put("prevTCCL", Thread.currentThread().getContextClassLoader());
-    Thread.currentThread().setContextClassLoader(custom);
+    ClassLoader instrumentationClassloader =
+        ctx.getStore(NAMESPACE).get(INSTRUMENTATION_CLASSLOADER, ClassLoader.class);
+    ctx.getStore(NAMESPACE)
+        .put(ORIGINAL_CLASSLOADER, Thread.currentThread().getContextClassLoader());
+    Thread.currentThread().setContextClassLoader(instrumentationClassloader);
   }
 
   @Override
   public void afterEach(ExtensionContext ctx) {
-    ClassLoader prev = ctx.getStore(NAMESPACE).remove("prevTCCL", ClassLoader.class);
+    ClassLoader prev = ctx.getStore(NAMESPACE).remove(ORIGINAL_CLASSLOADER, ClassLoader.class);
     if (prev != null) {
       Thread.currentThread().setContextClassLoader(prev);
     }
@@ -86,11 +114,7 @@ public final class SpockExtension
     }
   }
 
-  /**
-   * Class‑loader that <em>shadows</em> the test class, so that any re‑definitions via ByteBuddy
-   * operate on a loader‑private copy instead of the one used by the build toolʼs class‑path
-   * scanning. Delegation order = child‑first for the testʼs own package, parent‑first otherwise.
-   */
+  /** Run test classes in a classloader which loads test classes before delegating. */
   private static class InstrumentationClassLoader extends ClassLoader {
     private final ClassLoader parent;
     private final String shadowPrefix;
@@ -101,7 +125,7 @@ public final class SpockExtension
       this.shadowPrefix = shadowPrefix;
     }
 
-    /** Inject the bytes of {@code clazz} into <b>this</b> loader, producing a shadow copy. */
+    /** Forcefully inject the bytes of clazz into this classloader. */
     Class<?> shadow(Class<?> clazz) throws IOException {
       Class<?> loaded = findLoadedClass(clazz.getName());
       if (loaded != null && loaded.getClassLoader() == this) {
